move _UnicodeSafeStreamWrapper into a dedicated package for internal logging stuff as it's not for end-users; create internal/ package for items that should not be documented
