+++ pysys/baserunner.py
@@ -43,7 +43,7 @@ from pysys.basetest import BaseTest
 from pysys.process.user import ProcessUser
 from pysys.utils.logutils import BaseLogFormatter
 from pysys.utils.pycompat import *
-from pysys.utils.pycompat import _UnicodeSafeStreamWrapper
+from pysys.internal.initlogging import _UnicodeSafeStreamWrapper
 
 from pysys.writer import ConsoleSummaryResultsWriter, ConsoleProgressResultsWriter, BaseSummaryResultsWriter, BaseProgressResultsWriter
